style ongoing

+++ src/views/AdminArea.vue
@@ -1,61 +1,97 @@
 <template>
   <div class="admin-area-container">
-    <Header/>
-    <!-- <div class="login">
-      <p>Login</p>
-      <input type="email" v-model="admin.email" placeholder="email" />
-      <input type="password" v-model="admin.password" placeholder="password" />
-      <button v-on:click="loginAdmin">login</button>
-      <p>User: {{loggedIn}}</p>
-    </div> -->
+    <Header />
+    <div class="add-product-button">
+      <img :src="require('@/assets/img/add-circle-outline.svg')"
+      @click="toggleCreateMode"
+      />
+    </div>
     <div class="user-container" v-if="getCurrentUser != null">
       <p>User: {{getCurrentUser.name}}</p>
       <p>Role: {{getCurrentUser.role}}</p>
     </div>
-    <div class="create-product">
-      <p>Create Product</p>
-      <input type="text" v-model="product.title" placeholder="title" />
-      <input type="number" v-model="product.price" placeholder="price" />
-      <input type="text" v-model="product.shortDesc" placeholder="Short description" />
-      <textarea type="text" v-model="product.longDesc" placeholder="Long description"></textarea>
-      <select v-on:change="showImgCreate" v-model="product.imgFile">
-        <option value="hoodie-ash.png">hoodie-ash.png</option>
-        <option value="hoodie-fire.png">hoodie-fire.png</option>
-        <option value="hoodie-ocean.png">hoodie-ocean.png</option>
-        <option value="skateboard-generic.png">skateboard-generic.png</option>
-        <option value="skateboard-greta.png">skateboard-greta.png</option>
-        <option value="wheel-rocket.png">wheel-rocket.png</option>
-        <option value="wheel-spinner.png">wheel-spinner.png</option>
-        <option value="wheel-wave.png">wheel-wave.png</option>
-      </select>
-      <img v-if="displayImgCreate" v-bind:src="require('@/assets/img/' + product.imgFile)" />
-      <button @click="createProduct">Create</button>
+    <div class="create-product-container" v-if="activateCreateMode">
+      <div class="create-product">
+        <img class="exit" :src="require('@/assets/img/close-circle-outline.svg')"
+        @click="toggleCreateMode">
+        <div class="product-input-container">
+        <input type="text" v-model="product.title" placeholder="title" />
+        <input type="number" v-model="product.price" placeholder="price" />
+        <input type="text" v-model="product.shortDesc" placeholder="Short description" />
+        <textarea type="text" v-model="product.longDesc" placeholder="Long description"></textarea>
+        <button class="create-button" @click="createProduct">Create Product</button>
+        </div>
+        <div class="product-in-update-container">
+          <article class="product-in-update">
+
+          <img v-if="displayImgCreate" v-bind:src="require('@/assets/img/' + product.imgFile)" />
+          
+        <select v-on:change="showImgCreate" v-model="product.imgFile">
+          <option value="hoodie-ash.png">hoodie-ash.png</option>
+          <option value="hoodie-fire.png">hoodie-fire.png</option>
+          <option value="hoodie-ocean.png">hoodie-ocean.png</option>
+          <option value="skateboard-generic.png">skateboard-generic.png</option>
+          <option value="skateboard-greta.png">skateboard-greta.png</option>
+          <option value="wheel-rocket.png">wheel-rocket.png</option>
+          <option value="wheel-spinner.png">wheel-spinner.png</option>
+          <option value="wheel-wave.png">wheel-wave.png</option>
+        </select>
+        </article>
+        </div>
+      </div>
     </div>
-    <div class="updateProduct" v-if="activateUpdateMode">
-      <p>Update Product</p>
-      <input type="text" v-model="updateProductForm.title" placeholder="title" />
-      <input type="number" v-model="updateProductForm.price" placeholder="price" />
-      <input type="text" v-model="updateProductForm.shortDesc" placeholder="Short description" />
-      <textarea type="text" v-model="updateProductForm.longDesc" placeholder="Long description"></textarea>
-      <select v-model="updateProductForm.imgFile">
-        <option value="hoodie-ash.png">hoodie-ash.png</option>
-        <option value="hoodie-fire.png">hoodie-fire.png</option>
-        <option value="hoodie-ocean.png">hoodie-ocean.png</option>
-        <option value="skateboard-generic.png">skateboard-generic.png</option>
-        <option value="skateboard-greta.png">skateboard-greta.png</option>
-        <option value="wheel-rocket.png">wheel-rocket.png</option>
-        <option value="wheel-spinner.png">wheel-spinner.png</option>
-        <option value="wheel-wave.png">wheel-wave.png</option>
-      </select>
-      <img v-if="displayImgUpdate" v-bind:src="require('@/assets/img/' + updateProductForm.imgFile)" />
-      <button @click="updateProduct">send update</button>
+    <div class="update-product-container" v-if="activateUpdateMode">
+      <div class="update-product">
+        <img class="exit" :src="require('@/assets/img/close-circle-outline.svg')"
+        @click="toggleUpdateMode">
+        <div class="product-input-container">
+        <input type="text" v-model="updateProductForm.title" placeholder="title" />
+        <input type="number" v-model="updateProductForm.price" placeholder="price" />
+        <input type="text" v-model="updateProductForm.shortDesc" placeholder="Short description" />
+        <textarea type="text" v-model="updateProductForm.longDesc" placeholder="Long description"></textarea>
+        <button class="update-button" @click="updateProduct">Update Product</button>
+        </div>
+        <div class="product-in-update-container">
+          <article class="product-in-update">
+          <img
+          v-if="displayImgUpdate"
+          v-bind:src="require('@/assets/img/' + updateProductForm.imgFile)"
+          /> 
+        <select v-model="updateProductForm.imgFile">
+          <option value="hoodie-ash.png">hoodie-ash.png</option>
+          <option value="hoodie-fire.png">hoodie-fire.png</option>
+          <option value="hoodie-ocean.png">hoodie-ocean.png</option>
+          <option value="skateboard-generic.png">skateboard-generic.png</option>
+          <option value="skateboard-greta.png">skateboard-greta.png</option>
+          <option value="wheel-rocket.png">wheel-rocket.png</option>
+          <option value="wheel-spinner.png">wheel-spinner.png</option>
+          <option value="wheel-wave.png">wheel-wave.png</option>
+        </select>
+        </article>
+        </div>
+      </div>
     </div>
     <div class="product-container">
       <article class="product" v-for="product in getProductList" v-bind:key="product._id">
-        <p>{{product.title}} {{product._id}}</p>
-        <img v-bind:src="require('@/assets/img/' + product.imgFile)" />
-        <button @click="toggleUpdateMode(product)">update product</button>
-        <button @click="deleteProduct(product)">delete</button>
+        <div class="product-frame">
+          <div class="icons">
+            <img
+              :src="require('@/assets/img/create-outline.svg')"
+              @click="toggleUpdateMode(product)"
+            />
+            <img :src="require('@/assets/img/trash-outline.svg')" @click="deleteProduct(product)" />
+          </div>
+          <div class="product-img">
+            <img v-bind:src="require('@/assets/img/' + product.imgFile)" />
+          </div>
+        </div>
+        <article class="product-info">
+          <p>{{product.title}}</p>
+          <p>{{product.shortDesc}}</p>
+          <p>
+            <strong>{{product.price}} kr inkl moms</strong>
+          </p>
+        </article>
       </article>
     </div>
   </div>
@@ -87,7 +123,8 @@ export default {
       },
       activateUpdateMode: false,
       displayImgCreate: false,
-      displayImgUpdate: false
+      displayImgUpdate: false,
+      activateCreateMode: false
     };
   },
   computed: {
@@ -107,22 +144,34 @@ export default {
     },
     updateProduct() {
       this.$store.dispatch("updateProduct", this.updateProductForm);
+      this.activateUpdateMode = !this.activateUpdateMode;
     },
     deleteProduct(product) {
-      this.$store.dispatch("deleteProduct", product);
+      if (confirm("Are you sure?")) {
+        this.$store.dispatch("deleteProduct", product);
+      }
     },
     toggleUpdateMode(product) {
-      this.displayImgUpdate = true
+      if(this.activateCreateMode == true){
+        this.activateCreateMode = false
+      }
+      this.displayImgUpdate = true;
       this.updateProductForm = product;
       this.activateUpdateMode = !this.activateUpdateMode;
     },
     showImgCreate() {
-      this.displayImgCreate = true
+      this.displayImgCreate = true;
     },
+    toggleCreateMode() {
+      if(this.activateUpdateMode == true){
+        this.activateUpdateMode = false
+      }
+      this.activateCreateMode = !this.activateCreateMode;
+    }
   },
 
   created() {
-    this.$store.commit('restoreSession')
+    this.$store.commit("restoreSession");
   }
 };
 </script>
@@ -131,38 +180,150 @@ export default {
 .admin-area-container {
   max-width: 1000px;
   margin: 0 auto;
+  background: #f0f0f0;
 }
-.login {
-  border: 1px solid black;
+.product-container {
+  max-width: 800px;
+  margin: 0 auto;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+.product {
+}
+.product-frame {
+  -webkit-box-shadow: 1px 1px 5px 0px rgba(0, 0, 0, 0.75);
+  -moz-box-shadow: 1px 1px 5px 0px rgba(0, 0, 0, 0.75);
+  box-shadow: 1px 1px 5px 0px rgba(0, 0, 0, 0.75);
+  width: 20rem;
+  height: 15rem;
+  background: #ebebeb;
+  border-radius: 1rem;
+  padding: 1rem;
+  margin: 1rem;
+}
+.product-frame .icons {
+  margin: 0 0 1rem 0;
+  display: flex;
+  justify-content: space-between;
+}
+.product-frame .icons img {
+  display: flex;
+  justify-content: space-between;
+  width: 10%;
+}
+.product-frame .icons img:hover {
+  cursor: pointer;
+}
+.product-img img {
+  width: 50%;
+}
+.product-info p {
+  margin: 0;
+  text-align: left;
+  margin-left: 2rem;
+}
+.update-product-container {
+  max-width: 800px;
+  margin: 0 auto;
+}
+.update-product {
+  -webkit-box-shadow: 1px 1px 5px 0px rgba(0, 0, 0, 0.75);
+  -moz-box-shadow: 1px 1px 5px 0px rgba(0, 0, 0, 0.75);
+  box-shadow: 1px 1px 5px 0px rgba(0, 0, 0, 0.75);
+  width: 50rem;
+  height: 30rem;
+  background: #C4C4C4;
+  border-radius: 2rem;
+}
+.update-product {
+  position: relative;
+  display: flex;
+  justify-content: space-evenly;
+  align-items: center;
+}
+.update-product .exit {
+  padding: 1rem;
+  width: 5%;
+  position:absolute; top:0; right:0;
+}
+.update-product .exit:hover {
+  cursor: pointer;
+}
+.product-input-container {
+  padding: 1rem;
+  height: 21rem;
+  width: 20rem;
   display: flex;
   flex-direction: column;
-  align-items: center;
+  justify-content: space-around;
+}
+.product-input-container input {
+  height: 2rem;
+  font-size: 1rem;
+}
+.product-input-container textarea {
+  height: 6rem;
+  font-size: 1rem;
+}
+.product-in-update-container {
+  height: 21rem;
+  display: flex;
+}
+.product-in-update {
+  -webkit-box-shadow: 1px 1px 5px 0px rgba(0, 0, 0, 0.75);
+  -moz-box-shadow: 1px 1px 5px 0px rgba(0, 0, 0, 0.75);
+  box-shadow: 1px 1px 5px 0px rgba(0, 0, 0, 0.75);
+  width: 20rem;
+  height: 15rem;
+  background: #ebebeb;
+  border-radius: 1rem;
+  padding: 1rem;
+  margin: 1rem;
+}
+.product-in-update img {
+  height: 90%;
+  padding: 1rem;
+}
+.product-in-update select {
+  margin: 1rem;
+}
+.update-button {
+  padding: 1rem;
+  border-radius: 3rem;
+}
+.add-product-button img {
+  width: 5%;
+}
+.create-product-container {
+  max-width: 800px;
+  margin: 0 auto;
+}
+.create-product {
+  position: relative;
+  -webkit-box-shadow: 1px 1px 5px 0px rgba(0, 0, 0, 0.75);
+  -moz-box-shadow: 1px 1px 5px 0px rgba(0, 0, 0, 0.75);
+  box-shadow: 1px 1px 5px 0px rgba(0, 0, 0, 0.75);
+  width: 50rem;
+  height: 30rem;
+  background: #C4C4C4;
+  border-radius: 2rem;
+}
+.create-product .exit {
+  padding: 1rem;
+  width: 5%;
+  position:absolute; top:0; right:0;
+}
+.create-product .exit:hover {
+  cursor: pointer;
 }
 .create-product {
   display: flex;
-  flex-direction: column;
+  justify-content: space-evenly;
   align-items: center;
 }
-.create-product form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-.product-container {
-  display: flex;
-  justify-content: center;
-  flex-wrap: wrap;
-}
-.product {
-  width: 15rem;
-  border: 1px dotted red;
-}
-img {
-  width: 10%;
-}
-.updateProduct {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
+.create-button {
+  padding: 1rem;
+  border-radius: 3rem;
 }
 </style>
